+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed `typing_extensions` from runtime dependencies https://github.com/Textualize/rich/pull/3763
 - Live objects (including Progress) may now be nested https://github.com/Textualize/rich/pull/3768
+- Added padding property to Syntax which returns a tuple of four integers https://github.com/Textualize/rich/pull/3782
 
 ### Fixed
 
 - Fixed extraction of recursive exceptions https://github.com/Textualize/rich/pull/3772
+- Fixed padding applied to Syntax https://github.com/Textualize/rich/pull/3782
 
 ### Added